requirements.txt
streamlit
pypdf
python-pptx
scikit-learn
requests
